Created first "draft" of IBCF
Created trained IBCF model (IBCF_train) and saved the model (IBCF_model) as the training takes a while to run. Still a useful practice for predictions as we don't want to rerun models all the time unnecessarily.

The recommender prediction function has been built but is not yet working as it should. I've included it but it still needs work.

Minor update to master to include reference to IBCF model